requirements.txt
Flask==3.1.0 
Flask-Mail==0.10.0 
Flask-SQLAlchemy==3.1.1 
Flask-Migrate==4.1.0 
Flask-Login==0.6.3 
Flask-WTF==1.2.2 
Flask-Bcrypt==1.0.1 
gunicorn==23.0.0 
requests==2.32.3 
beautifulsoup4==4.12.3 
Pillow==10.3.0 
python-dotenv==1.0.1 
SQLAlchemy==2.0.36 
mysql-connector-python==9.2.0 
bcrypt==4.3.0 
itsdangerous==2.2.0